+++ frontend/src/types/spaceFooter.type.ts
